requirements.txt
poe-api
discord